+++ go.mod
@@ -2,15 +2,13 @@ module github.com/imiller31/protoc-gen-tsql
 
 go 1.20
 
-require github.com/lyft/protoc-gen-star/v2 v2.0.3
+require (
+	github.com/golang/protobuf v1.5.2
+	github.com/lyft/protoc-gen-star/v2 v2.0.3
+	google.golang.org/protobuf v1.30.0
+)
 
 require (
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/lyft/protoc-gen-star v0.6.2 // indirect
 	github.com/spf13/afero v1.3.3 // indirect
-	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
-	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
 	golang.org/x/text v0.3.7 // indirect
-	golang.org/x/tools v0.1.12 // indirect
-	google.golang.org/protobuf v1.30.0 // indirect
 )
