+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/imiller31/protoc-gen-tsql/pkg/tsqlschemafy"
 	pgs "github.com/lyft/protoc-gen-star/v2"
 )
 
@@ -8,6 +9,6 @@ func main() {
 	pgs.Init(
 		pgs.DebugEnv("DEBUG"),
 	).RegisterModule(
-		Tsqlfy(),
+		tsqlschemafy.TsqlSchemafy(),
 	).Render()
 }
