+++ pkg/tsqlschemafy/types.go
@@ -1,0 +1,13 @@
+package tsqlschemafy
+
+type TsqlfySchema struct {
+	TableName string
+	Columns   []TsqlColumns
+
+	PrimaryKeyColumns []string
+}
+
+type TsqlColumns struct {
+	Name    string
+	SqlType string
+}
