+++ pkg/tsqlschemafy/visitor.go
@@ -1,79 +1,11 @@
-package main
+package tsqlschemafy
 
 import (
-	"bytes"
-	"embed"
 	"strconv"
-	"text/template"
 
 	"github.com/imiller31/protoc-gen-tsql/proto-ext-tsql/tsql_options"
 	pgs "github.com/lyft/protoc-gen-star/v2"
 )
-
-//go:embed templates/create_table.tmpl
-var createTableTmpl embed.FS
-
-var funcs = template.FuncMap{
-	"sub": func(a, b int) int {
-		return a - b
-	},
-}
-
-type TsqlfyModule struct {
-	*pgs.ModuleBase
-}
-
-type TsqlfySchema struct {
-	TableName string
-	Columns   []TsqlColumns
-
-	PrimaryKeyColumns []string
-}
-
-type TsqlColumns struct {
-	Name    string
-	SqlType string
-}
-
-func Tsqlfy() *TsqlfyModule { return &TsqlfyModule{ModuleBase: &pgs.ModuleBase{}} }
-
-func (p *TsqlfyModule) Name() string { return "tsqlfy" }
-
-func (p *TsqlfyModule) Execute(targets map[string]pgs.File, packages map[string]pgs.Package) []pgs.Artifact {
-	for _, f := range targets {
-		p.generateSchema(f)
-	}
-
-	return p.Artifacts()
-}
-
-func (p *TsqlfyModule) generateSchema(f pgs.File) {
-	p.Push(f.Name().String())
-	defer p.Pop()
-
-	schema := TsqlfySchema{}
-
-	v := initTsqlfyVisitor(p.BuildContext, &schema)
-	p.CheckErr(pgs.Walk(v, f), "unable to construct TSQL schema")
-
-	p.BuildContext.Logf("TSQL Schema:\n%s", schema)
-	t := template.New("create_table.tmpl").Funcs(funcs)
-	t, err := t.ParseFS(createTableTmpl, "templates/create_table.tmpl")
-	if err != nil {
-		p.Fail("unable to parse template", err)
-	}
-	var out bytes.Buffer
-	if err := t.Execute(&out, schema); err != nil {
-		p.Fail("unable to execute template", err)
-	}
-
-	p.Logf("TSQL:\n%s", out.String())
-
-	p.AddGeneratorFile(
-		f.InputPath().SetExt("_schema.sql").String(),
-		out.String(),
-	)
-}
 
 type TsqlfyVisitor struct {
 	pgs.Visitor
